Make the behaviour for selecting top level cards match the comments.
The code made ordering assumptions that weren't correct for selections.

This resulted in us adding expansion level items to the selection after
selecting the top level item. This led to strange and unpredictable results,
such as unexpectedly deleting cards from the set entirely when setting the
number to less than the current value as we'd remove cards for both the
top-level and expansion level selections.

We now check for top-levels items correctly. The behaviour may not be entirely
intuitive, since we ignore the expansion level selection entirely, but it is
predictable and consistent.



git-svn-id: http://svn.code.sf.net/p/sutekh/code/trunk@3739 a4bb362d-5410-0410-a86c-8388c1e7b32c